Refactor realm properties
(cherry picked from commit 1b52634a899cb4ac9d9deb60326647b5e3a32e72)

+++ Source/ACE.Database/ShardDatabase.cs
@@ -500,7 +500,7 @@ namespace ACE.Database
             return staticObjects;
         }
 
-        public List<Biota> GetDynamicObjectsByLandblock(ushort landblockId, byte instance)
+        public List<Biota> GetDynamicObjectsByLandblock(ushort landblockId, uint instance)
         {
             var dynamics = new List<Biota>();
